% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormPredictionplot.R
\name{convert_arg}
\alias{convert_arg}
\title{Function to convert any input to string vector}
\usage{
convert_arg(x)
}
\arguments{
\item{x}{string, object name or vector of theese}
}
\value{
vector of strings
}
\description{
Function to convert any input to string vector
}
