% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{educ_data}
\alias{educ_data}
\title{Morfometrics of children}
\format{A data frame with 11032 rows and 10 variables:
\describe{
  \item{educ_3}{highest obtained educational level}
  \item{Rural}{location of school (rural or urban)}
  \item{sex}{geneder of the child)}
  \item{max_SEP_3}{highest parental proffesion}
  \item{n_siblings}{number of children in the family}
  \item{cran_rzs}{cranial volume (residuals to age an birth date)}
  \item{height_rzs}{height (residuals to age an birth date)}
  \item{FW_rzs}{face width (residuals to age an birth date)}
  \item{YOB}{year of birth}
  \item{YOBc}{centered but not scaled year of birth (YOB)}


}}
\usage{
educ_data
}
\description{
A dataset containing the standardized residuals of Estonain schoolchildren.
}
\keyword{datasets}
