% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{forestplot.default}
\alias{forestplot.default}
\title{Draw forestplot from an orm model summary}
\usage{
forestplot.default(summary_obj, return_ggplots = FALSE,
  plot.widths = c(0.5, 0.5), title = "Odds Ratio", ...)
}
\arguments{
\item{summary_obj}{result of a \code{summary} command on
\code{\link[rms]{orm}}  model}

\item{plot.widths}{the relative widths of the left and right plot
should be a vector (\code{c()})  with 2 elements that sum to 1 defaults to
equal widths}

\item{title}{the tile row of the drawn plot}

\item{...}{Arguments passed on to \code{oddstable_graph}
\describe{
  \item{digits}{the number of siginficants digits to display}
  \item{theme}{the desired ggplot2 theme}
  \item{header}{names of the table columns}
  \item{row.names.y}{new names for the variable rows}
}}
}
\description{
Function to get aligned table and plot of the Odds ratio
}
\details{
#' @param return_ggplots if \code{TRUE} the fuction returns 2 ggplot objects
 in a list instead of drawing a tablegrid
}
