% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{join_ggplots}
\alias{join_ggplots}
\title{Join two ggplot objects side by side}
\usage{
join_ggplots(leftplot, rightplot, plot.widths = c(0.5, 0.5),
  title = "Odds Ratio", ...)
}
\arguments{
\item{plot.widths}{the relative widths of the left and right plot
should be a vector (\code{c()})  with 2 elements that sum to 1 defaults to
equal widths}

\item{title}{the tile row of the drawn plot}
}
\description{
Function to get aligned table of two ggplot objects
}
