% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{oddstable}
\alias{oddstable}
\title{Get row names from odd an values form even columns}
\usage{
oddstable(x)
}
\arguments{
\item{x}{a \code{matrix} with even number of rows}
}
\description{
Get row names from odd an values form even columns
}
