% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{oddstable_graph}
\alias{oddstable_graph}
\title{Make a ggplot table}
\usage{
oddstable_graph(
  x,
  digits = 3,
  theme = ggplot2::theme_get(),
  header = NULL,
  row.names.y = NULL
)
}
\arguments{
\item{x}{a \code{matrix} or a \code{data.frame}}

\item{digits}{the number of significant digits to display}

\item{theme}{the desired ggplot2 theme}

\item{header}{names of the table columns}

\item{row.names.y}{new names for the variable rows}
}
\description{
Function to get a ggplot table from a matrix
}
