% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{orm_graph}
\alias{orm_graph}
\title{Make a ggplot figure}
\usage{
orm_graph(
  x,
  theme = ggplot2::theme_get(),
  header = NULL,
  row.names.y = NULL,
  shape = 19,
  limits = NULL,
  breaks = c(0.5, 1, 1.5, 2, 3, 4)
)
}
\arguments{
\item{x}{a \code{matrix} or a \code{data.frame}}

\item{theme}{the desired ggplot2 theme}

\item{header}{names of the table columns}

\item{row.names.y}{new names for the variable rows}

\item{shape}{point shape, see \code{\link[ggplot2]{aes_linetype_size_shape}}}

\item{limits}{the x axis limits as a vector, see also:
\code{\link[ggplot2]{scale_continuous}}}

\item{breaks}{the x axis breaks as a vector,help see also:
\code{\link[ggplot2]{scale_continuous}}}
}
\description{
Function to get a ggplot figure from a matrix x
}
