% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{orm_summary.default}
\alias{orm_summary.default}
\title{Get an extended rms summary object}
\usage{
orm_summary.default(object, ...)
}
\description{
This fuctions adds a \code{summary.orm} class attribute to a
\code{summary.rms} output to facilitate the automatic plotting
 of orm summary into a ggplot forestplot
}
\details{
@param object a \code{\link[rms]{orm}}  model object
}
