% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{plot.summary.orm}
\alias{plot.summary.orm}
\title{Forest Plot of an orm model summary}
\usage{
\method{summary}{orm}(summary_obj, return_ggplots = FALSE, ..., digits,
  theme, header, row.names.y)
}
\arguments{
\item{summary_obj}{result of a \code{summary} command on
\code{\link[rms]{orm}}  model}

\item{...}{Arguments passed on to \code{join_ggplots}
\describe{
  \item{plot.widths}{the relative widths of the left and right plot
should be a vector (\code{c()})  with 2 elements that sum to 1 defaults to
equal widths}
  \item{title}{the tile row of the drawn plot}
}}

\item{digits}{the number of siginficants digits to display}

\item{theme}{the desired ggplot2 theme}

\item{header}{names of the table columns}

\item{row.names.y}{new names for the variable rows}
}
\value{
Two \code{ggplot} plot objects or a \code{\link[gtable]{gtable}}
}
\description{
Convinience function to create a plot of the \code{\link[rms]{orm}}  model
summary. For further customising the plots use \code{return_ggplots = TRUE}
This will create 2 \code{ggplot2} objects that can be joined with the
\code{\link{join_ggplots}} commands.
}
