% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{plot.summary.rms}
\alias{plot.summary.rms}
\title{Forest Plot of an rms model summary}
\usage{
\method{plot}{summary.rms}(x, ...)
}
\arguments{
\item{x}{result of a \code{summary} command on
\code{\link[rms]{orm}}  model ie a summary.orm class object}

\item{...}{Arguments passed on to \code{forestplot}
\describe{
  \item{return_ggplots}{if \code{TRUE} the fuction returns 2 ggplot objects
in a list instead of drawing a tablegrid}
  \item{x}{result of a \code{summary} command on
\code{\link[rms]{orm}}  model ie a summary.orm class object}
  \item{plot.widths}{the relative widths of the left and right plot
should be a vector (\code{c()})  with 2 elements that sum to 1 defaults to
equal widths}
  \item{title}{the tile row of the drawn plot}
}}
}
\description{
Convinience function to create a plot of the \code{\link[rms]{orm}} or
\code{\link[rms]{lrm}} model summary. For further customising the plots
 use \code{return_ggplots = TRUE}
This will create 2 \code{ggplot2} objects that can be joined with the
\code{\link{join_ggplots}} commands.
}
