% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{summary.orm}
\alias{summary.orm}
\title{Get an extended rms summary object}
\usage{
\method{summary}{orm}(object, ...)
}
\description{
This fuctions adds a \code{summary.orm} class attribute to a
\code{summary.rms} output to facilitate the automatic plotting
 of orm summary into a ggplot forestplot
}
