% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ormForestplot.R
\name{plot.summary.orm}
\alias{plot.summary.orm}
\alias{plot}
\title{Forest Plot of an orm model summary}
\usage{
\method{plot}{summary.orm}(x, ...)
}
\arguments{
\item{x}{result of a \code{summary} command on
\code{\link[rms]{orm}}  model ie a summary.orm class object}

\item{...}{paremeters passed to \code{\link{forestplot}}}
}
\description{
Convinience function to create a plot of the \code{\link[rms]{orm}}  model
summary. For further customising the plots use \code{return_ggplots = TRUE}
This will create 2 \code{ggplot2} objects that can be joined with the
\code{\link{join_ggplots}} commands.
}
